pinit/requirements.txt
Pillow==9.4.0
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.1.7
django-crispy-forms==1.13.0
idna==3.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8